fix 單身更新時單別、單號、序號不允許變更

+++ grails-app/controllers/foodpaint/OutSrcPurchaseSheetController.groovy
@@ -92,6 +92,13 @@ class OutSrcPurchaseSheetController {
         
         def outSrcPurchaseSheet = OutSrcPurchaseSheet.get(params.id)
 
+        //單別、單號一旦建立不允許變更
+        if(params.typeName != outPurchaseSheet.typeName || params.name != outPurchaseSheet.name){
+            render (contentType: 'application/json') {
+                [success: false,message:message(code: 'sheet.typeName.name.not.allowed.change')]
+            }
+            return
+        }
         def result
         try {
             result = domainService.delete(outSrcPurchaseSheet)
